Split out main.jsx and used global: process.env
